Tests: Added tests for StorageManager. Code refactoring.
