+++ src/storagemanager.js
